path_syncer: move syncBlocks and helpers out of conflictResolver
Issue: KBFS-2064
